Consolidate lib/ into src/ and change output to dist/
Closes #153.
