+++ .eslintignore
@@ -1,8 +1,4 @@
 coverage/
+dist/
 test/web-platform-tests/
-lib/URL.js
-lib/URLSearchParams.js
-lib/url-state-machine.js
-lib/urlencoded.js
-lib/utils.js
 live-viewer/whatwg-url.js
