+++ .gitignore
@@ -9,10 +9,6 @@ deploy_key.pub
 # project specific
 coverage/
 test/web-platform-tests/
-lib/URL.js
-lib/URLSearchParams.js
-lib/url-state-machine.js
-lib/urlencoded.js
-lib/utils.js
-live-viewer/whatwg-url.js
+dist/
 out/
+live-viewer/whatwg-url.js
