+++ scripts/deploy-live-viewer.sh
@@ -4,7 +4,7 @@ set -o nounset
 set -o pipefail
 cd "$(dirname "$0")"
 
-yarn build-live-viewer
+npm run build-live-viewer
 
 mkdir -p ../out
 cp -r ../live-viewer/* ../out # exclude files starting with . by using /*
